Move naming tests into the naming test module
